modified as per comment
