+++ template/client/store/getters.js
@@ -1,0 +1,11 @@
+// @flow
+
+import type {
+  StateType
+} from '~/types'
+
+export const getters = {
+  hasComments(state: StateType): boolean {
+    return Boolean(state.comments && state.comments.length > 0)
+  }
+}
