+++ template/client/store/mutations.js
@@ -1,0 +1,46 @@
+// @flow
+
+import type {
+  CommentType,
+  CommentPayloadType,
+  StateType,
+  RawCommentType
+} from '~/types'
+
+const SET_COMMENTS = 'root/SET_COMMENTS'
+const UPDATE_RATING = 'root/UPDATE_RATING'
+
+export const mutations = {
+  [SET_COMMENTS]: (
+    state: StateType, comments: Array<RawCommentType>
+  ) => {
+    const updatedComments: Array<CommentType> = []
+
+    for (const comment of comments.slice(0, 10)) {
+      const newOne = {
+        ...comment,
+        rating: 0 // try to comment out this line
+      }
+      updatedComments.push(newOne)
+    }
+
+    state.comments = updatedComments
+  },
+
+  [UPDATE_RATING]: (
+    state: StateType, {
+      commentId,
+      delta
+    }: CommentPayloadType
+  ) => {
+    if (!state.comments) return
+
+    const commentIndex = state.comments.findIndex((c: CommentType) => {
+      return c.id === commentId
+    })
+
+    if (!state.comments || !state.comments[commentIndex]) return
+
+    state.comments[commentIndex].rating += delta
+  }
+}
