+++ template/client/store/state.js
@@ -1,0 +1,5 @@
+// @flow
+
+export const state = () => {
+  comments: null
+}
