--- template/client/types/mutations.js
@@ -1,4 +1,0 @@
-// @flow
-
-export const SET_COMMENTS = 'root/SET_COMMENTS'
-export const UPDATE_RATING = 'root/UPDATE_RATING'
